Refactor UnityTools into partials, add new test suites
Refactored UnityTools.cs into multiple partial class files for modularity and maintainability, each targeting a specific area of Unity tool functionality. Added three new comprehensive NUnit test files—UnityToolsNewToolsTests.cs, UnityToolsRecipeTests.cs, and UnityYamlWriterTests.cs—covering direct tool operations, orchestration recipes, and YAML serialization. All tests use NSubstitute for mocking and provide broad coverage of the Unity MCP tooling layer. No functional changes to tool or test logic; this is a pure structural and test coverage improvement.
